Migrate to Eigen.
Removes the Series class, and with it, a lot of the previous code. I
learned what I wanted to learn about move semantics, operator
overloading, iterators, storage and strides, so there would not be any
benefit to labouriously maintaining my own algebra package.

Other changes:
- Merges `examples/` into `assignments/`. Compiles the assignment code
with CMake.
- Rewrites the spectrogram (assignment 2) and `AudioFile` to use
`Eigen::ArrayXX` instead of `std::vector`. This makes the code much
cleaner, such as finding the min and max gains in the spectrogram.
- Cleans up the a1 code to clearly demonstrate all required
functionality.
- Improves the CSV reader to ignore quotation marks which are ubiquitous
in the sample datasets.